Version 1.0a!
This is the initial version of the game. I still have a long way to go to get to what I envisioned when I wrote down the concept of this game, but the core is there. I can't wait to continue on with this project and improve on the game.
The game is fully functional and infinitely re-playable.